requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-allauth==0.44.0
django-bootstrap-modal-forms==2.2.0
django-dajax==0.9.2
django-dajaxice==0.7
django-filter==2.4.0
filelock==3.0.12
identify==2.2.0
idna==2.10
jinja-to-js==3.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
nodeenv==1.5.0
oauthlib==3.1.0
Pillow==8.2.0
pre-commit==2.11.1
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.3
